NativeIO: Add quota to async Blink interface
This CL adds a quota management system to NativeIO’s Blink
implementation. Before writing to a file (write, setLength), the
application must explicitly ask for capacity. When deleting or shrinking
files, the freed-up space is given back to the application. Capacity is
allocated per Javascript execution context.

This CL implements the support for the asynchronous methods of NativeIO.

Blink requests capacity from the browser process. Implementing the
browser side will be done in a followup CL. For now, the browser grants
any request up to 8 GiB.

The design doc for NativeIO's quota system is here:
https://docs.google.com/document/d/1wUrtCOsyH3qGwKuqLhV9AJD-bDAjAzfPL5r1GT8H4IY

Bug: 1137788
Change-Id: I9f528373b1e7f5bf45fa8070964cdc3f225d1657